Shut up a bunch of GCC warnings
Mostly int<->ptr conversions on x86_64

git-svn-id: http://vapoursynth.googlecode.com/svn/trunk@46 f9120d27-2007-6f97-8312-0f4ebfa7710f

+++ src/core/vsresize.c
@@ -241,7 +241,7 @@ static void VS_CC resizeCreate(const VSMap *in, VSMap *out, void *userData, VSCo
     d.lsrch = 0;
     d.lsrcw = 0;
     d.node = 0;
-    d.flags = (int)userData;
+    d.flags = (intptr_t)userData;
     d.node = vsapi->propGetNode(in, "clip", 0, 0);
     d.vi = *vsapi->getVideoInfo(d.node);
     dstwidth = vsapi->propGetInt(in, "width", 0, &err);
